+++ nav2_graceful_controller/doc/trajectories.png
