+++ nav2_graceful_controller/test/CMakeLists.txt
@@ -1,14 +1,14 @@
 find_package(nav2_controller REQUIRED)
 
-# Tests for Graceful Motion Controller
-ament_add_gtest(test_graceful_motion_controller
-  test_graceful_motion_controller.cpp
+# Tests for Graceful Controller
+ament_add_gtest(test_graceful_controller
+  test_graceful_controller.cpp
 )
-ament_target_dependencies(test_graceful_motion_controller
+ament_target_dependencies(test_graceful_controller
   ${dependencies}
   nav2_controller
 )
-target_link_libraries(test_graceful_motion_controller
+target_link_libraries(test_graceful_controller
   ${library_name}
 )
