--- nav2_graceful_motion_controller/graceful_controller_plugin.xml
@@ -1,7 +1,0 @@
-<class_libraries>
-    <library path="nav2_graceful_motion_controller">
-        <class type="nav2_graceful_motion_controller::GracefulMotionController" base_class_type="nav2_core::Controller">
-            <description>Graceful motion controller for Nav2</description>
-        </class>
-    </library>
-</class_libraries>
